+++ src/components/date/__spec__.js
@@ -18,7 +18,7 @@ describe('Date', () => {
   describe('intialize', () => {
     it('sets the intial internal state', () => {
       expect(instance.state.open).toBeFalsy();
-      expect(instance.state.viewDate).toBeNull();
+      expect(instance.state.datePickerValue).toBeNull();
       expect(instance.state.visibleValue).toEqual(today);
     });
   });
@@ -144,10 +144,20 @@ describe('Date', () => {
     });
 
     it('calls set state to open the view and set the view date', () => {
-      expect(instance.setState).toHaveBeenCalledWith({
-        open: true,
-        viewDate: instance.props.defaultValue
-      })
+      expect(instance.setState).toHaveBeenCalledWith({ open: true })
+    });
+
+    describe('date validity', () => {
+      describe('when a valid date', () => {
+        it('calls set state setting the datePickerValue to be the valid date', () => {
+          instance = TestUtils.renderIntoDocument(
+            <Date name='date' value='2015/01/01' label='Date' />
+          );
+          spyOn(instance, 'setState');
+          instance.openDatePicker();
+          expect(instance.setState).toHaveBeenCalledWith({ datePickerValue: '2015/01/01' });
+        });
+      });
     });
   });
 
@@ -193,7 +203,7 @@ describe('Date', () => {
       expect(instance.emitOnChangeCallback).toHaveBeenCalledWith(hiddenToday);
       expect(instance.setState).toHaveBeenCalledWith({
         visibleValue: today,
-        viewDate: hiddenToday
+        datePickerValue: hiddenToday
       });
     });
 
@@ -206,16 +216,7 @@ describe('Date', () => {
         instance.handleVisibleInputChange({ target: { value: date } })
         expect(instance.setState).toHaveBeenCalledWith({
           visibleValue: date,
-          viewDate: hiddenDate
-        });
-      });
-
-      it('accepts the format MMM DD YY', () => {
-        let date = moment().add(noOfDays, 'days').format('MMM DD YY');
-        instance.handleVisibleInputChange({ target: { value: date } })
-        expect(instance.setState).toHaveBeenCalledWith({
-          visibleValue: date,
-          viewDate: hiddenDate
+          datePickerValue: hiddenDate
         });
       });
 
@@ -224,7 +225,7 @@ describe('Date', () => {
         instance.handleVisibleInputChange({ target: { value: date } })
         expect(instance.setState).toHaveBeenCalledWith({
           visibleValue: date,
-          viewDate: hiddenToday
+          datePickerValue: hiddenToday
         });
       });
 
@@ -233,7 +234,7 @@ describe('Date', () => {
         instance.handleVisibleInputChange({ target: { value: date } })
         expect(instance.setState).toHaveBeenCalledWith({
           visibleValue: date,
-          viewDate: hiddenDate
+          datePickerValue: hiddenDate
         });
       });
 
@@ -242,7 +243,7 @@ describe('Date', () => {
         instance.handleVisibleInputChange({ target: { value: date } })
         expect(instance.setState).toHaveBeenCalledWith({
           visibleValue: date,
-          viewDate: hiddenDate
+          datePickerValue: hiddenDate
         });
       });
     });
@@ -371,10 +372,10 @@ describe('Date', () => {
   });
 
   describe('handleViewDateChange', () => {
-    it('sets the state of the viewDate', () => {
+    it('sets the state of the datePickerValue', () => {
       spyOn(instance, 'setState');
       instance.handleViewDateChange(123);
-      expect(instance.setState).toHaveBeenCalledWith({ viewDate: 123 });
+      expect(instance.setState).toHaveBeenCalledWith({ datePickerValue: 123 });
     });
   });